del cast

+++ 307_YOLOv7/post_process_gen_tools/make_final_batch_nums_final_class_nums_final_box_nums.py
@@ -14,7 +14,7 @@ class Model(nn.Module):
     def forward(self, x):
         batch_nums = x[:, 0:1].to(torch.float32) # batch number
         class_nums = x[:, 1:2].to(torch.float32) # class ids
-        box_nums = x[:, [0,2]].to(torch.float32) # batch number + box number
+        box_nums = x[:, [0,2]] # batch number + box number
         return batch_nums, class_nums, box_nums
